Extract network utilities
